requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.2
rasterio==1.3.9
pandas==2.1.3
